Added: Recording support for conditional recording

+++ src/test/kotlin/io/sip3/salto/ce/rtpr/RtprSessionTest.kt
@@ -174,8 +174,8 @@ class RtprSessionTest {
         assertEquals(RTPR_1.startedAt + RTPR_1.duration, session.terminatedAt)
 
         assertTrue(session.codecNames.contains(RTPR_1.codecName))
-        assertEquals(RTPR_1.mos, session.mos)
-        assertEquals(RTPR_1.rFactor, session.rFactor)
+        assertEquals(RTPR_1.mos.toDouble(), session.mos)
+        assertEquals(RTPR_1.rFactor.toDouble(), session.rFactor)
     }
 
     @Test
@@ -196,7 +196,7 @@ class RtprSessionTest {
         assertEquals(RTPR_2.startedAt + RTPR_2.duration, session.terminatedAt)
 
         assertTrue(session.codecNames.contains(RTPR_1.codecName))
-        assertEquals((RTPR_1.rFactor + RTPR_2.rFactor) / 2, session.rFactor)
+        assertEquals((RTPR_1.rFactor + RTPR_2.rFactor) / 2.toDouble(), session.rFactor)
     }
 
     @Test
@@ -222,7 +222,7 @@ class RtprSessionTest {
         assertEquals(RTPR_2.startedAt + RTPR_2.duration, session.terminatedAt)
 
         assertTrue(session.codecNames.contains(RTPR_1.codecName))
-        assertEquals((RTPR_1.rFactor + RTPR_2.rFactor) / 2, session.rFactor)
+        assertEquals((RTPR_1.rFactor + RTPR_2.rFactor) / 2.toDouble(), session.rFactor)
     }
 
     @Test
@@ -250,7 +250,7 @@ class RtprSessionTest {
         assertEquals(RTPR_1_RTCP.startedAt + RTPR_1_RTCP.duration, session.terminatedAt)
 
         assertTrue(session.codecNames.contains(RTPR_1_RTCP.codecName))
-        assertEquals(RTPR_1_RTCP.mos, session.mos)
-        assertEquals(RTPR_1_RTCP.rFactor, session.rFactor)
+        assertEquals(RTPR_1_RTCP.mos.toDouble(), session.mos)
+        assertEquals(RTPR_1_RTCP.rFactor.toDouble(), session.rFactor)
     }
 }
